keeping around older Extras packages, we can end up with some pretty pathological multilib cases around noarch packages.
Having both x86_64 and noarch of the same package installed is
non-sensical.  But we want to allow switching from native -> noarch and
back.  So, let's take into account that noarch is a third case beyond
base arch and compat arch.

With that, we get to where there are four basic cases of newest packages
1) No noarch packages.  This ends up with what we have today where you
install the base arch and the compat arch if available
2) Only noarch packages.  This also ends up like present
3) Noarch package that's newer than the basearch package.  This means
that we want the noarch package and not the basearch or compat-arch
4) Noarch package that's older than the basearch package.  This means we
want the basearch and compat-arch if it exists

rh #189998